+++ lib/ansible/module_utils/oneandone.py
@@ -192,6 +192,18 @@ def get_vpn(oneandone_conn, vpn, full_object=False):
             return _vpn['id']
 
 
+def get_public_ip(oneandone_conn, public_ip, full_object=False):
+    """
+    Validates that the public ip exists by ID or a name.
+    Returns the public ip if one was found.
+    """
+    for _public_ip in oneandone_conn.list_public_ips(per_page=1000):
+        if public_ip in (_public_ip['id'], _public_ip['ip']):
+            if full_object:
+                return _public_ip
+            return _public_ip['id']
+
+
 def wait_for_resource_creation_completion(oneandone_conn,
                                           resource_type,
                                           resource_id,
